+++ src/behaviour.js
@@ -1,7 +1,0 @@
-const ROTATION_SPEED = 0.002;
-
-export default {
-  rotate(object, delta) {
-    object.setRotation(object.rotation + ROTATION_SPEED * delta);
-  },
-};